+++ src/main/java/com/ebicep/warlords/abilties/HolyRadianceCrusader.java
@@ -23,7 +23,7 @@ public class HolyRadianceCrusader extends AbstractHolyRadianceBase {
     private int markDuration = 8;
     private int energyPerSecond = 6;
 
-    public HolyRadianceCrusader(float minDamageHeal, float maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier) {
+    public HolyRadianceCrusader(float minDamageHeal, float maxDamageHeal, float cooldown, float energyCost, float critChance, float critMultiplier) {
         super("Holy Radiance", minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier, 6);
     }
